requirements.txt
Django==3.0
django-environ==0.4.5
djangorestframework==3.11.1
markdown==3.2.2
django-filter==2.3.0
Pillow==7.2.0
gunicorn==20.0.4
django-heroku==0.3.1
dj-database-url==0.5.0
whitenoise==5.2.0
psycopg2==2.8.5
drf-nested-routers==0.91
Faker==4.1.2